# public/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/count": {
      "get": {
        "operationId": "UserCurdController_count",
        "summary": "Get user count",
        "description": "Retrieve the total count of users based on query parameters.",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User count retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number",
                  "example": 10
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/options": {
      "get": {
        "operationId": "UserCurdController_options",
        "summary": "Get user options",
        "description": "Fetch user options based on query parameters.",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User options retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": [
                    {
                      "label": "user.name",
                      "value": "user._id",
                      "imageUrl": "user.imageUrl",
                      "description": "user.description"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user": {
      "get": {
        "operationId": "UserCurdController_findAll",
        "summary": "Get all users",
        "description": "Retrieve a paginated list of all users.",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UserCurdController_create",
        "summary": "Create a new user",
        "description": "Create a new user in the system.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserCurdController_findById",
        "summary": "Get user by ID",
        "description": "Retrieve a user by their unique ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserCurdController_update",
        "summary": "Update user by ID",
        "description": "Update an existing user by their ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "User verification failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserCurdController_delete",
        "summary": "Delete user by ID",
        "description": "Delete a user by their unique ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Boilerplate v1 API",
    "description": "Basic docs for API in version 1",
    "version": "1.0",
    "contact": {
      "name": "Shirsho ",
      "url": "https://marketron.com",
      "email": "hniqbal01@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid request parameters"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-11-22T04:10:35.729Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "shirsho"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "shirsho@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone",
            "example": "1234567890"
          },
          "imageUrl": {
            "type": "string",
            "description": "User image url",
            "example": "image.jpg"
          },
          "description": {
            "type": "string",
            "description": "User description",
            "example": "I am a developer"
          },
          "role": {
            "description": "User role",
            "example": [
              "USER"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolePermission": {
            "type": "string",
            "description": "User role permission",
            "example": "user"
          },
          "isverified": {
            "type": "boolean",
            "description": "User is verified",
            "example": false
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "imageUrl",
          "description",
          "role",
          "rolePermission",
          "isverified"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho"
          },
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "phone": {
            "type": "string",
            "example": "01534115844"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "imageUrl": {
            "type": "string",
            "example": "imageUrl"
          },
          "description": {
            "type": "string",
            "example": "description"
          },
          "provider": {
            "type": "string",
            "example": "Google"
          },
          "providerId": {
            "type": "string",
            "example": "local"
          },
          "role": {
            "example": "user",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolePermission": {
            "type": "string",
            "example": "user"
          }
        }
      },
...
      "UserPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho"
          },
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "phone": {
            "type": "string",
            "example": "01534115844"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "imageUrl": {
            "type": "string",
            "example": "imageUrl"
          },
          "description": {
            "type": "string",
            "example": "description"
          },
          "provider": {
            "type": "string",
            "example": "Google"
          },
          "providerId": {
            "type": "string",
            "example": "local"
          },
          "role": {
            "example": "user",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolePermission": {
            "type": "string",
            "example": "user"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "phone",
          "status",
          "imageUrl",
          "description",
          "provider",
          "providerId",
          "role"
        ]
      },
...
    }
  }
}